requirements.txt
python-binance>=1.0.15
psycopg2-binary>=2.9.1
